NUM-167 Rename AmqpRequestMessageWithProperties to AmqpQueueMessageWithProperties and move it back to akka
