+++ amqp-driver/src/main/scala/com/nummulus/amqp/driver/akka/AkkaRabbitConsumer.scala
@@ -3,7 +3,6 @@ package com.nummulus.amqp.driver.akka
 import java.nio.charset.StandardCharsets
 
 import com.nummulus.amqp.driver.MessageProperties
-import com.nummulus.amqp.driver.api.provider.AmqpRequestMessageWithProperties
 import com.rabbitmq.client.AMQP
 import com.rabbitmq.client.Channel
 import com.rabbitmq.client.DefaultConsumer
@@ -19,6 +18,6 @@ private[driver] class AkkaRabbitConsumer(channel: Channel, actor: ActorRef) exte
      val message = new String(body, StandardCharsets.UTF_8)
      val deliveryTag = envelope.getDeliveryTag
      
-     actor ! AmqpRequestMessageWithProperties(message, MessageProperties(properties), deliveryTag)
+     actor ! AmqpQueueMessageWithProperties(message, MessageProperties(properties), deliveryTag)
    }
 }
