+++ amqp-driver/src/main/scala/com/nummulus/amqp/driver/akka/AmqpQueueMessageWithProperties.scala
@@ -1,4 +1,4 @@
-package com.nummulus.amqp.driver.api.provider
+package com.nummulus.amqp.driver.akka
 
 import com.nummulus.amqp.driver.MessageProperties
 
@@ -9,4 +9,4 @@ import com.nummulus.amqp.driver.MessageProperties
  * @param properties message properties
  * @param deliveryTag amqp message identifier
  */
-private[driver] case class AmqpRequestMessageWithProperties(body: String, properties: MessageProperties, deliveryTag: Long)
+private[driver] case class AmqpQueueMessageWithProperties(body: String, properties: MessageProperties, deliveryTag: Long)
